requirements.txt
pyserial >= 3.0.1
click >= 6.2